21f1000692_previous/ticket_app/backend/requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
passlib==1.7.4
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
aniso8601==9.0.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
passlib==1.7.4
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
aniso8601==9.0.1
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
passlib==1.7.4
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
